Update SP2013-PurgeVersionHistory-KeepVersions.ps1
In SharePoint 2013, you can enable version history and see all of the different versions of the documents. I had a request to first purge version history older than 6 months and for any versions that remain, they would want to keep only 20 versions. Since there were many documents in the library, I created a script to run once and to meet this requirement.