Added Script to TransferZipFilesInCertainDateRange
Requirement was to get archive folders with zip files for the year of 2018 and to copy the contents in the zip file and place them in another directory with the same name. If any contents in the zip files are also in the first half of 2018, they should be deleted.